PSWTM-49 function patchglobalRoomConfig, globale Room Config aendern

+++ BuildingMonitoringFunctionsapp/src/classes/RoomConfig.cs
@@ -10,23 +10,23 @@ namespace BuildingMonitoringFunctionsapp
         //  private static string queryID = "select [targetTemper],[targetHumid],[updateRate],[upperToleranceTemper],[lowerToleranceTemper],[upperToleranceHumid]," +
         //                              "[lowerToleranceHumid] from roomConfig  " +
         //                               "where id=(select configId from room where id=@roomID)";
-        //private static string queryID = "select * from roomConfig  " +
+        // private static string queryID = "select * from roomConfig  " +
         //                                   "where id=@roomId";
         public int roomId { get; set; }
 
-        public double targetTemper { get; set; }
+        public float targetTemper { get; set; }
 
-        public double targetHumid { get; set; }
+        public float targetHumid { get; set; }
 
-        public double updateRate { get; set; }
+        public float updateRate { get; set; }
 
-        public double upperToleranceTemper { get; set; }
+        public float upperToleranceTemper { get; set; }
 
-        public double lowerToleranceTemper { get; set; }
+        public float lowerToleranceTemper { get; set; }
 
-        public double upperToleranceHumid { get; set; }
+        public float upperToleranceHumid { get; set; }
 
-        public double lowerToleranceHumid { get; set; }
+        public float lowerToleranceHumid { get; set; }
     }
 }
